extract pdf gen logic to a helper
